Deler opp ArbeidOgInntekt i minde komponenter
