+++ packages/v2/backend/package.json
@@ -23,7 +23,7 @@
   "dependencies": {
     "@navikt/k9-klage-typescript-client": "2.0.20250218125133",
     "@navikt/k9-sak-typescript-client": "2.0.20250320132234",
-    "@navikt/ung-sak-typescript-client": "0.2.20250328135442"
+    "@navikt/ung-sak-typescript-client": "0.2.20250407085651"
   },
   "devDependencies": {
     "@tanstack/eslint-plugin-query": "^5.71.5"
