+++ packages/v2/gui/src/prosess/ung-beregning/aksjonspunktArbeidOgInntekt.module.css
@@ -1,0 +1,3 @@
+.personIcon {
+  flex-shrink: 0;
+}
