+++ packages/v2/gui/src/prosess/ung-beregning/aksjonspunktArbeidOgInntekt.module.d.css.ts
@@ -1,0 +1,5 @@
+declare const styles: {
+  readonly "personIcon": string;
+};
+export = styles;
+
