+++ packages/v2/gui/src/prosess/ung-beregning/arbeidOgInntekt.module.css
@@ -18,21 +18,3 @@
 .checkmarkIcon {
   color: var(--a-surface-success);
 }
-
-.personIcon {
-  flex-shrink: 0;
-}
-
-.sum {
-  align-items: center;
-  background-color: #417da0;
-  border-radius: 4px;
-  color: white;
-  display: flex;
-  justify-content: center;
-  padding: 5px;
-}
-
-.sumLabel {
-  color: #156389;
-}
