+++ packages/v2/gui/src/prosess/ung-beregning/arbeidOgInntekt.module.d.css.ts
@@ -4,9 +4,6 @@ declare const styles: {
   readonly "firstDataCell": string;
   readonly "firstHeaderCell": string;
   readonly "lastRow": string;
-  readonly "personIcon": string;
-  readonly "sum": string;
-  readonly "sumLabel": string;
 };
 export = styles;
