+++ packages/v2/gui/src/prosess/ung-beregning/detaljerOmInntekt.module.css
@@ -1,0 +1,13 @@
+.sum {
+  align-items: center;
+  background-color: #417da0;
+  border-radius: 4px;
+  color: white;
+  display: flex;
+  justify-content: center;
+  padding: 5px;
+}
+
+.sumLabel {
+  color: #156389;
+}
