+++ packages/v2/gui/src/prosess/ung-beregning/detaljerOmInntekt.module.d.css.ts
@@ -1,0 +1,6 @@
+declare const styles: {
+  readonly "sum": string;
+  readonly "sumLabel": string;
+};
+export = styles;
+
